src/improvement_notifier/pyproject.toml
[project]
name = "improvement-notifier"
version = "0.1.0"
description = "Discord hook to remind about incomplete improvements"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
